[REFACTOR]기본프로필변경
