+++ src/pages/CrewBoardCreate.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { createCrewRecruitmentPost } from "@/api/supabase-api/crewRecruitmentPost";
 import { getCurrentUser } from "@/api/supabase-api/userInfo";
-import Baseball from "@/assets/icons/baseball.svg";
+import Baseball from "@/assets/icons/default_profile_xl.svg";
 import Calendar from "@/assets/icons/calendar.svg";
 import DropdownSelect from "@/components/common/DropdownSelect.vue";
 import Loading from "@/components/common/Loading.vue";
@@ -264,7 +264,8 @@ const handleDropdownToggle = (key) => {
               type="text"
               v-model="content"
               class="w-full p-0 outline-none resize-none text-4 placeholder-gray01 placeholder-4 bg-white01"
-              placeholder=" 직관 크루 모집글을 작성해보세요" />
+              placeholder=" 직관 크루 모집글을 작성해보세요"
+            />
           </div>
         </div>
         <div class="flex flex-col gap-[30px]">
@@ -279,11 +280,13 @@ const handleDropdownToggle = (key) => {
               :options="recruitOptions"
               part="모집 상태"
               :isOpen="openDropdown === 'recruitStatus'"
-              @toggle="handleDropdownToggle('recruitStatus')" />
+              @toggle="handleDropdownToggle('recruitStatus')"
+            />
           </div>
           <div class="flex justify-between">
             <div
-              class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+              class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+            >
               <div class="w-[100px]">
                 <span
                   class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -291,24 +294,28 @@ const handleDropdownToggle = (key) => {
                 >
               </div>
               <div
-                class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+                class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+              >
                 <div class="relative w-full">
                   <button
                     @click="isDatePickerOpen = !isDatePickerOpen"
-                    class="flex items-center justify-between w-full h-[40px] px-[15px] bg-white02 text-black01 text-[18px] rounded-[8px] border">
+                    class="flex items-center justify-between w-full h-[40px] px-[15px] bg-white02 text-black01 text-[18px] rounded-[8px] border"
+                  >
                     <span>{{ formattedGameDate || "날짜를 선택하세요" }}</span>
                     <img :src="Calendar" class="w-[18px] h-[18px]" />
                   </button>
                   <div
                     v-if="isDatePickerOpen"
-                    class="absolute z-10 mt-2 left-[350px] top-[-10px]">
+                    class="absolute z-10 mt-2 left-[350px] top-[-10px]"
+                  >
                     <VDatePicker v-model="gameDateStatus" mode="single" />
                   </div>
                 </div>
               </div>
             </div>
             <div
-              class="flex justify-between items-center h-[40px] gap-2 w-[425px]">
+              class="flex justify-between items-center h-[40px] gap-2 w-[425px]"
+            >
               <div class="w-[100px]">
                 <span
                   class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -321,14 +328,16 @@ const handleDropdownToggle = (key) => {
                   :options="peopleNumOptions"
                   part="인원"
                   :isOpen="openDropdown === 'peopleNum'"
-                  @toggle="handleDropdownToggle('peopleNum')" />
+                  @toggle="handleDropdownToggle('peopleNum')"
+                />
 
                 <DropdownSelect
                   v-model:selectedOption="peopleStatus"
                   :options="peopleStatusOptions"
                   part="인원 상태"
                   :isOpen="openDropdown === 'peopleStatus'"
-                  @toggle="handleDropdownToggle('peopleStatus')" />
+                  @toggle="handleDropdownToggle('peopleStatus')"
+                />
               </div>
             </div>
           </div>
@@ -342,7 +351,8 @@ const handleDropdownToggle = (key) => {
             </div>
             <div class="flex justify-between">
               <div
-                class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+                class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+              >
                 <div class="w-[100px]">
                   <span
                     class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -354,10 +364,12 @@ const handleDropdownToggle = (key) => {
                   :options="myTeamOptions"
                   part="응원팀"
                   :isOpen="openDropdown === 'myTeam'"
-                  @toggle="handleDropdownToggle('myTeam')" />
+                  @toggle="handleDropdownToggle('myTeam')"
+                />
               </div>
               <div
-                class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+                class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+              >
                 <div class="w-[100px]">
                   <span
                     class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -369,7 +381,8 @@ const handleDropdownToggle = (key) => {
                   :options="stadiumOptions"
                   part="경기 장소"
                   :isOpen="openDropdown === 'stadium'"
-                  @toggle="handleDropdownToggle('stadium')" />
+                  @toggle="handleDropdownToggle('stadium')"
+                />
               </div>
             </div>
           </div>
@@ -383,7 +396,8 @@ const handleDropdownToggle = (key) => {
             </div>
             <div class="flex justify-between">
               <div
-                class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+                class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+              >
                 <div class="w-[100px]">
                   <span
                     class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -395,10 +409,12 @@ const handleDropdownToggle = (key) => {
                   :options="myGenderOptions"
                   part="작성자 성별"
                   :isOpen="openDropdown === 'myGender'"
-                  @toggle="handleDropdownToggle('myGender')" />
+                  @toggle="handleDropdownToggle('myGender')"
+                />
               </div>
               <div
-                class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+                class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+              >
                 <div class="w-[100px]">
                   <span
                     class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -410,13 +426,15 @@ const handleDropdownToggle = (key) => {
                   :options="myAgeOptions"
                   part="작성자 연령"
                   :isOpen="openDropdown === 'myAge'"
-                  @toggle="handleDropdownToggle('myAge')" />
+                  @toggle="handleDropdownToggle('myAge')"
+                />
               </div>
             </div>
           </div>
           <div class="flex justify-between">
             <div
-              class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+              class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+            >
               <div class="w-[100px]">
                 <span
                   class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -429,10 +447,12 @@ const handleDropdownToggle = (key) => {
                 :disabled="isCrewGenderDisabled"
                 part="크루 성별"
                 :isOpen="openDropdown === 'crewGender'"
-                @toggle="handleDropdownToggle('crewGender')" />
+                @toggle="handleDropdownToggle('crewGender')"
+              />
             </div>
             <div
-              class="flex justify-between items-center h-[40px] w-[425px] gap-2">
+              class="flex justify-between items-center h-[40px] w-[425px] gap-2"
+            >
               <div class="w-[100px]">
                 <span
                   class="text-black01 text-bold text-[18px] whitespace-nowrap"
@@ -444,7 +464,8 @@ const handleDropdownToggle = (key) => {
                 :options="crewAgeOptions"
                 part="크루 연령"
                 :isOpen="openDropdown === 'crewAge'"
-                @toggle="handleDropdownToggle('crewAge')" />
+                @toggle="handleDropdownToggle('crewAge')"
+              />
             </div>
           </div>
         </div>
